ai-career-coach/requirements.txt
streamlit
google-generativeai
pdfplumber
python-docx
sentence-transformers
faiss-cpu
python-dotenv
